Updated README file and folder structure
