+++ README.txt
@@ -8,45 +8,62 @@ README.txt
 
 alternatebitprotocol.doc
 
-atomics [This folder contains the header files]
-	receiverCadmium.hpp
-	senderCadmium.hpp
-	subnetCadmium.hpp
+data [This folder contains the input and output files for the simulator]
+	input [This folder contains the input files]
+		input_abp_0.txt
+		input_abp_1.txt
+	output [This folder contains the output files]
+		abp_output.txt
 
-data_structures [This folder contains the data structures used in the project]
-	message.hpp
-	message.cpp
+doc [This folder contains the documentation]
+	alternatebitprot.pdf
+	Cadmium_Documentation_Ubuntu.pdf
+	Cadmium_Windows.pdf
 
-test [This folder the unit test for the different include files]
-	receiver [This folder contains the unit test of the receiver]
+include [This folder contains the various header files used in the project]
+	atomics [This folder contains the atomics header files ]
+		receiverCadmium.hpp
+		senderCadmium.hpp
+		subnetCadmium.hpp
+	data_structures [This folder contains the data structures headers]
+		message.hpp	
+	vendor [This folder contains vendor (third party) header files]
+		iestream.hpp
+		NDTime.hpp
+
+lib [This folder contains the third party library files]
+	cadmium [This folder contains the cadmium (third party) library files]
+	DESTimes [This folder contains the DESTimes (third party) library files]
+	
+src [This folder contains the source files used for the simualator] 
+	data_structures [This folder contains the source files from data_structures]
+		message.cpp
+	top_model [This folder contains the source files from top_model]
 		main.cpp
-		makefile
-		receiver_input_test.txt
-		receiver_test_output.txt
-	sender [This folder contains the unit test of the sender]
-		main.cpp
-		makefile
-		sender_input_test_ack_In.txt
-		sender_input_test_control_In.txt
-		sender_test_output.txt
-	subnet [This folder contains the unit test of the subnet]
-		main.cpp
-		makefile
-		subnet_input_test.txt
-		subnet_test_output.txt
 
-vendor [This folder contains 3rd party header files needed in the project
-	iestream.hpp
-	NDTime.hpp
+test [This folder contains files realted to the unit tests]
+	data [This folder contains the input and output files for the unit tests]
+		sender [This folder contains some data for sender testing and simulator outputs]
+			sender_input_test_ack_In.txt
+			sender_input_test_control_In.txt
+			sender_test_output.txt
+		receiver [This folder contains some data for receiver testing and simulator outputs] 
+			receiver_input_test.txt
+			receiver_test_output.txt
+		subnet [This folder contains some data for subnet testing and simulator outputs]
+			subnet_input_test.txt
+			subnet_test_output.txt
 
-top_model [This folder contains source code for the Alternate Bit Protocol simulator
-some data for testing and some simulator outputs]	
-	main.cpp
-	makefile
-	input_abp_0.txt
-	input_abp_1.txt
-	abp_output_0.txt
-	abp_output_1.txt
+	include [This folder will include header files related to unit tests]
+
+	src [This folder contains the source code files for the unit tests]
+		sender [This folder contains source code for the sender component]
+			main.cpp
+		receiver [This folder contains source code for the receiver component] 
+			main.cpp
+		subnet [This folder contains source code for the subnet component]
+			main.cpp
+makefile
 	
 /*************/
 /****STEPS****/
@@ -60,47 +77,28 @@ some data for testing and some simulator outputs]
 	Example: INCLUDECADMIUM=-I lib/cadmium/include
 
 2 - Run the simulator
-		1 - Open the terminal. Press in your keyboard Ctrl+Alt+t
-		2 - Set the command prompt in the AlternateBitProtocol folder. To do so, type in the terminal the path to this folder.
-			Example: cd ../../AlternateBitProtocol
-		3 - To compile the project, type in the terminal:
-			make clean; make all
-		4 - To run the simulation, go the folder with executable files (i.e., bin folder). To do so, type in the terminal the path to this folder.
-			Example: cd bin
-		5 - In the terminal type "./NAME_OF_THE_COMPILED_FILE NAME_OF_THE_INPUT_FILE". For this test you need to type:
-			./ABP ../data/input/input_abp_1.txt
-		6 - To check the output of the simulation, open "../data/output/abp_output.txt"
-		7 - To execute the simulator with different inputs
-			7.1. Create new .txt files with the same structure as input_abp_0.txt or input_abp_1.txt
-			7.2. Run the simulator using the instructions in step 4
-			7.3. If you want to keep the output, rename abp_output.txt. To do so, type in the terminal: "mv abp_output.txt NEW_NAME"
-			Example: mv abp_output.txt abp_output_0.txt
-
-=====
-
-2 - Run the unit tests
-	2.1. Run subnet test
-		1 - Open the terminal. Press in your keyboard Ctrl+Alt+t
-		2 - Set the command prompt in the test/subnet folder. To do so, type in the terminal the path to this folder.
-			Example: cd Documents/AlternateBitProtocol/test/subnet
-		3 - To compile the test, type in the terminal:
-			make clean; make all
-		4 - To run the test, type in the terminal "./NAME_OF_THE_COMPILED_FILE". For this specific test you need to type:
-			./SUBNET
-		5 - To check the output of the test, open  "subnet_test_output.txt"
-	2.2. To run receiver and sender tests, the steps are analogous to 2.1
-			
-3 - Run the simulator
 	1 - Open the terminal. Press in your keyboard Ctrl+Alt+t
-	2 - Set the command prompt in the top_model folder. To do so, type in the terminal the path to this folder.
-		Example: cd ../../top_model
+	2 - Set the command prompt in the AlternateBitProtocol folder. To do so, type in the terminal the path to this folder.
+		Example: cd ../../AlternateBitProtocol
 	3 - To compile the project, type in the terminal:
 		make clean; make all
-	4 - To run the simulation, type in the terminal "./NAME_OF_THE_COMPILED_FILE NAME_OF_THE_INPUT_FILE". For this test you need to type:
-		./ABP input_abp_1.txt
-	5 - To check the output of the simulation, open  "abp_output.txt"
-	6 - To execute the simulator with different inputs
-		6.1. Create new .txt files with the same structure as input_abp_0.txt or input_abp_1.txt
-		6.2. Run the simulator using the instructions in step 4
-		6.3. If you want to keep the output, rename abp_output.txt. To do so, type in the terminal: "mv abp_output.txt NEW_NAME"
-			Example: mv abp_output.txt abp_output_0.txt
+	4 - To run the simulation, go the folder with executable files (i.e., bin folder). To do so, type in the terminal the path to this folder.
+		Example: cd bin
+	5 - In the terminal type "./NAME_OF_THE_COMPILED_FILE NAME_OF_THE_INPUT_FILE". For this test you need to type:
+		./ABP ../data/input/input_abp_1.txt
+	6 - To check the output of the simulation, open "../data/output/abp_output.txt"
+	7 - To execute the simulator with different inputs
+		7.1. Create new .txt files with the same structure as input_abp_0.txt or input_abp_1.txt
+		7.2. Run the simulator using the instructions in step 5
+		7.3. If you want to keep the output, rename abp_output.txt. To do so, type in the terminal: "mv abp_output.txt NEW_NAME"
+		Example: mv ../data/output/abp_output.txt ../data/output/abp_output_0.txt
+
+3 - Run the unit tests
+	3.1. Run subnet test
+		1 - Open the terminal. Press in your keyboard Ctrl+Alt+t
+		2 - To run the unit tests, go the folder with executable files (i.e., bin folder). To do so, type in the terminal the path to this folder.
+			Example: cd ../alternatebitprotocolsimulator/bin
+		3 - To run the test, type in the terminal "./NAME_OF_THE_COMPILED_FILE". For this specific test you need to type:
+			./SUBNET_TEST
+		4 - To check the output of the test, open ../test/data/subnet/subnet_test_output.txt
+	3.2. To run receiver and sender tests, the steps are analogous to 2.1
